move some rpc functionality to room
